backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-magic==0.4.27
python-dotenv==1.0.1

